utils/importer/requirements.txt
# ETL Process Dependencies
pandas>=1.3.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
uuid>=1.30